+++ frontend/src/Components/Pages/Patients/PatientPage.tsx
@@ -1,5 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { useEffect } from "react";
+import { confirmAlert } from "react-confirm-alert";
 import { useDispatch, useSelector } from "react-redux";
 import { useLocation, useNavigate } from "react-router-dom";
 import { deleteUser, UserData } from "../../../Api/IdentityData";
@@ -14,6 +15,7 @@ import {
   DangerButton,
   FormButtonContainer,
   PrimaryButton,
+  SecondaryButton,
 } from "../../../Styles/Common/Buttons";
 import { doctorAdditionalInf } from "../../../Styles/Doctors/DoctorStyles";
 import {
@@ -21,6 +23,13 @@ import {
   patientTitle,
 } from "../../../Styles/Patient/PatientPageStyles";
 import { Page } from "../../General/Page";
+
+import "react-confirm-alert/src/react-confirm-alert.css";
+import {
+  confirmAlertContainer,
+  confirmAlertMessage,
+  confirmAlertTitle,
+} from "../../../Styles/Common/OtherComponents";
 
 type Props = {
   user: UserData;
@@ -45,20 +54,46 @@ export const PatientPage = ({ user }: Props) => {
         dispatch(signOutUserAction(location.pathname));
       dispatch(gotPatientAction(result.data));
     };
-    doGetPatient(user.id);
+    doGetPatient(loginedUser?.id);
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [loginedUser]);
 
   const editClickHandler = () => {
     navigate("/patients/" + patient?.id + "/edit");
   };
 
   const deleteClickHandler = async () => {
-    let result = await deleteUser(
-      user?.userName === undefined ? "" : user?.userName,
-      userToken === undefined ? "" : userToken,
-    );
-    if (result) dispatch(signOutUserAction("/"));
+    confirmAlert({
+      customUI: ({ onClose }) => {
+        return (
+          <div css={confirmAlertContainer}>
+            <h1 css={confirmAlertTitle}>Вы уверены?</h1>
+            <p css={confirmAlertMessage}>Это действие необратимо</p>
+            <FormButtonContainer className="row justify-content-around">
+              <SecondaryButton
+                className="btn btn-primary col-4 row"
+                onClick={onClose}
+              >
+                Нет
+              </SecondaryButton>
+              <DangerButton
+                className="btn btn-primary col-7 row"
+                onClick={async () => {
+                  let result = await deleteUser(
+                    user.userName === undefined ? "" : user.userName,
+                    userToken === undefined ? "" : userToken,
+                  );
+                  if (result) dispatch(signOutUserAction("/"));
+                  onClose();
+                }}
+              >
+                Да, удалить!
+              </DangerButton>
+            </FormButtonContainer>
+          </div>
+        );
+      },
+    });
   };
 
   return (
